Prompt for creation of dev entry.

+++ src/panoptes/pocs/utils/cli/mount.py
@@ -210,35 +210,41 @@ def setup_mount(
                     command_set = 'v310' if ra_fw >= 210101 and dec_fw >= 210101 else 'v250'
                     print(f'Suggested command set: {command_set}')
                     
-                    # Get info for writing udev entry.
-                    try:
-                        udev_str = (
-                            f'ACTION=="add", '
-                            f'SUBSYSTEM=="tty", '
-                            f'ATTRS{{idVendor}}=="{port.vid:04x}", '
-                            f'ATTRS{{idProduct}}=="{port.pid:04x}", '
-                        )
-                        if port.serial_number is not None:
-                            udev_str += f'ATTRS{{serial}}=="{port.serial_number}", '
-                        
-                        # The name we want it known by.    
-                        udev_str += f'SYMLINK+="ioptron"'
-                        
-                        udev_fn = Path('91-panoptes.rules')
-                        with udev_fn.open('w') as f:
-                            f.write(udev_str)
-                        
-                        print(f'Wrote udev entry to [green]{udev_fn}[/green].')
-                        print('Run the following command and then reboot for changes to take effect:')
-                        print(f'\t[green]cat {udev_fn} | sudo tee /etc/udev/rules.d/{udev_fn}[/green]')
-                    except Exception:
-                        pass
+                    write_port = port.device
+                    
+                    if typer.confirm('Do you want to make a udev entry?'):
+                        print('Creating udev entry for device')
+                        # Get info for writing udev entry.
+                        try:
+                            udev_str = (
+                                f'ACTION=="add", '
+                                f'SUBSYSTEM=="tty", '
+                                f'ATTRS{{idVendor}}=="{port.vid:04x}", '
+                                f'ATTRS{{idProduct}}=="{port.pid:04x}", '
+                            )
+                            if port.serial_number is not None:
+                                udev_str += f'ATTRS{{serial}}=="{port.serial_number}", '
+                            
+                            # The name we want it known by.
+                            udev_str += f'SYMLINK+="ioptron"'
+                            
+                            udev_fn = Path('91-panoptes.rules')
+                            with udev_fn.open('w') as f:
+                                f.write(udev_str)
+                            
+                            write_port = '/dev/ioptron'
+                            
+                            print(f'Wrote udev entry to [green]{udev_fn}[/green].')
+                            print('Run the following command and then reboot for changes to take effect:')
+                            print(f'\t[green]cat {udev_fn} | sudo tee /etc/udev/rules.d/{udev_fn}[/green]')
+                        except Exception:
+                            pass
                     
                     # Confirm the user wants to update the config.
                     if typer.confirm('Do you want to update the config?'):
                         print('Updating config.')
                         set_config('mount.brand', 'ioptron')
-                        set_config('mount.serial.port', '/dev/ioptron')
+                        set_config('mount.serial.port', write_port)
                         set_config('mount.serial.baudrate', baudrate)
                         set_config('mount.model', mount_type.name.lower())
                         set_config('mount.driver', f'panoptes.pocs.mount.ioptron.{mount_type.name.lower()}')
